+++ console/src/metadata/reducer.ts
@@ -1,6 +1,12 @@
 import { MetadataActions } from './actions';
-import { QueryCollection, HasuraMetadataV3, InheritedRole } from './types';
-import { allowedQueriesCollection } from './utils';
+import { HasuraMetadataV3, CollectionName, InheritedRole } from './types';
+import { setAllowedQueries } from './utils';
+
+export type AllowedQueriesCollection = {
+  name: string;
+  query: string;
+  collection: CollectionName;
+};
 
 type MetadataState = {
   metadataObject: null | HasuraMetadataV3;
@@ -8,7 +14,7 @@ type MetadataState = {
   loading: boolean;
   inconsistentObjects: any[];
   ongoingRequest: boolean; // deprecate
-  allowedQueries: QueryCollection[];
+  allowedQueries: AllowedQueriesCollection[];
   inheritedRoles: InheritedRole[];
 };
 
@@ -31,10 +37,10 @@ export const metadataReducer = (
       return {
         ...state,
         metadataObject: action.data,
-        allowedQueries:
-          action.data?.query_collections?.find(
-            query => query.name === allowedQueriesCollection
-          )?.definition.queries || [],
+        allowedQueries: setAllowedQueries(
+          action.data?.query_collections,
+          action.data?.allowlist
+        ),
         inheritedRoles: action.data?.inherited_roles,
         loading: false,
         error: null,