+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - server: fix action custom types failing to parse when mutually recursive
 - server: fix MSSQL table name descriptions
 - console: allow editing rest endpoints queries and misc ui improvements
+- console: display collection names and queries from all collections in allowlist
 - cli: match ordering of keys in project metadata files with server metadata
 
 ## v2.0.0-alpha.5